비밀번호 변경 창에서 이메일 확인 없애고, 필요없는 class 지움
