+++ lib/screen/profile/setting_profile.screen.dart
@@ -1,4 +1,7 @@
 import 'package:dating/controller/setting_controller.dart';
+import 'package:dating/controller/setting_password_controller.dart';
+import 'package:dating/data/repository/setting_password.repository.dart';
+import 'package:dating/data/service/setting_password_service.dart';
 import 'package:dating/screen/profile/account_information_screen.dart';
 import 'package:dating/screen/profile/blocked_account_screen.dart';
 import 'package:dating/style/icon_shape.dart';
@@ -64,7 +67,13 @@ class _SettingAccountScreenState extends State<SettingAccountScreen> {
           ),
           IconButton(
             onPressed: () {
-              Get.to(() => const AccountInformationScreen());
+              Get.to(() => const AccountInformationScreen(),
+                  binding: BindingsBuilder(() {
+                Get.put(SettingPasswordController(
+                    settingPasswordService: SettingPasswordService(
+                        settingPasswordRepository:
+                            SettingPasswordRepository())));
+              }));
             },
             icon: IconShape.iconArrowForward,
           ),
